Move skypilot function to sky.py
