+++ trainite/cli/init.py
@@ -210,18 +210,6 @@ def _recreation_command(config: "Init", project_name: str) -> str:
     return " ".join(cmd_parts)
 
 
-def get_skypilot_dependency() -> str:
-    """Get the canonical skypilot dependency string from Trainite's pyproject.toml."""
-    toml_path = PACKAGE_ROOT / "pyproject.toml"
-    if not toml_path.exists():
-        toml_path = PROJECT_ROOT / "pyproject.toml"
-    if toml_path.exists():
-        _, other_deps = parse_dependencies(toml_path)
-        if "skypilot" in other_deps:
-            return other_deps["skypilot"]
-    return "skypilot"
-
-
 def _build_templates(
     model_specs: Sequence[ModelSpec],
     dataset_spec: DatasetSpec,
